+++ src/Connection/PDOConnection.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * This file is part of Phuria SQL Builder package.
+ * This file is part of UnderQuery package.
  *
  * Copyright (c) 2016 Beniamin Jonatan Šimko
  *
@@ -32,93 +32,20 @@ class PDOConnection implements ConnectionInterface
     }
 
     /**
+     * @return \PDO
+     */
+    public function getWrappedConnection()
+    {
+        return $this->wrappedConnection;
+    }
+
+    /**
      * @param $SQL
      *
      * @return \PDOStatement
      */
-    private function prepareStatement($SQL)
+    public function prepareStatement($SQL)
     {
         return $this->wrappedConnection->prepare($SQL);
-    }
-
-    /**
-     * @param string                    $SQL
-     * @param QueryParameterInterface[] $parameters
-     *
-     * @return \PDOStatement
-     */
-    private function getExecutedStatement($SQL, array $parameters = [])
-    {
-        $preparedStmt = $this->prepareStatement($SQL);
-
-        foreach ($parameters as $parameter) {
-            $preparedStmt->bindValue($parameter->getName(), $parameter->getValue());
-        }
-
-        $preparedStmt->execute();
-
-        return $preparedStmt;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function execute($SQL, array $parameters = [])
-    {
-        $stmt = $this->getExecutedStatement($SQL, $parameters);
-
-        return $stmt->rowCount();
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function fetchScalar($SQL, array $parameters = [])
-    {
-        $stmt = $this->getExecutedStatement($SQL, $parameters);
-
-        if (0 < $stmt->rowCount()) {
-            return $stmt->fetch(\PDO::FETCH_COLUMN);
-        }
-
-        return null;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function fetchRow($SQL, array $parameters = [])
-    {
-        $stmt = $this->getExecutedStatement($SQL, $parameters);
-
-        if (0 < $stmt->rowCount()) {
-            return $stmt->fetch(\PDO::FETCH_ASSOC);
-        }
-
-        return [];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function fetchAll($SQL, array $parameters = [])
-    {
-        $stmt = $this->getExecutedStatement($SQL, $parameters);
-
-        if (0 < $stmt->rowCount()) {
-            return $stmt->fetchAll(\PDO::FETCH_ASSOC);
-        }
-
-        return [];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function rowCount($SQL, array $parameters = [])
-    {
-        $stmt = $this->getExecutedStatement($SQL, $parameters);
-
-        return $stmt->rowCount() ?: 0;
     }
 }