+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 Unreleased
 ---------
 
+### Changed
++ now Query contains only SQL and parameters
+
 0.5.0 [2016-10-26]
 -----------------